+++ nls/de/strings.js
@@ -1,8 +1,11 @@
 define({
     BEAUTIFY_ON_SAVE: 'Beautify beim Speichern',
+    BEAUTIFY_ON_SAVE_DESC: 'Aktiviert Beautify beim Speichern',
     UNSUPPORTED_TITLE: 'Sprache nicht unterstützt',
     UNSUPPORTED_MESSAGE: 'Diese Sprache ist nicht unterstützt.\n Verfügbare Sprachen sind JavaScript, JSON, HTML, XML, SVG, HTML in PHP, Embedded JavaScript, Handlebars, CSS, SCSS, und LESS.',
 
+    SASS_PREF_NAME: 'Pfad zum sass-convert Programm',
+    SASS_PREF_DESC: 'Absoluter Pfad zum sass-convert Programm',
     SASS_FORMAT: 'Fehler beim formatieren der SASS Datei',
-    SASS_ERROR: 'Bitte gib einen Pfad zum SASS-Convert program an.'
+    SASS_ERROR: 'Bitte gib einen Pfad zum SASS-Convert Programm an.'
 });