+++ nls/root/strings.js
@@ -1,9 +1,12 @@
 define({
     BEAUTIFY: 'Beautify',
     BEAUTIFY_ON_SAVE: 'Beautify on Save',
+    BEAUTIFY_ON_SAVE_DESC: 'true to Beautify file on Save',
     UNSUPPORTED_TITLE: 'Unsupported Language',
     UNSUPPORTED_MESSAGE: 'This language is not supported.\n Supported languages are JavaScript, JSON, HTML, XML, SVG, HTML in PHP, Embedded JavaScript, Handlebars, CSS, SCSS, and LESS.',
 
+    SASS_PREF_NAME: 'Path to sass-convert',
+    SASS_PREF_DESC: 'Absolute path to sass-convert',
     SASS_FORMAT: 'An error occurred formatting the SASS file',
     SASS_ERROR: 'You need to provide a path to the sass-convert program'
 });